valkyrie.plugins: Add Plugin manager.
